Work on xdebug enables but not success
In progress User creates a command
and adds in workers' table if none exists or adds a workspace.

Set current workspace query interactor.

Move roles domain.

In progress oauth

In progress Media and Storage domains

OAUTH GOOGLE IN PROGRESS
