+++ src/Media/Domain/Repository/SuperMediaRepositoryInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Media\Domain\Repository;
+
+interface SuperMediaRepositoryInterface
+{
+
+}
