+++ src/Media/Infrastructure/Database/ORM/SuperMedia.orm.xml
@@ -1,0 +1,40 @@
+<doctrine-mapping xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <mapped-superclass
+        name="App\Media\Domain\Entity\SuperMedia"
+        repository-class="App\Media\Infrastructure\Database\Repository\SuperMediaRepository"
+    >
+        <discriminator-column name="type" type="string"/>
+
+        <discriminator-map>
+            <discriminator-mapping value="document" class="App\Media\Domain\Entity\Document"/>
+            <discriminator-mapping value="archive" class="App\Media\Domain\Entity\Archive"/>
+        </discriminator-map>
+
+        <id name="id" type="string" length="26">
+            <generator strategy="NONE"/>
+        </id>
+
+        <field name="name" length="191"/>
+        <field name="originalName" column="original_name" length="500"/>
+        <field name="extension" length="7"/>
+        <field name="size" type="smallint"/>
+        <field name="fullPath" column="full_path" length="191"/>
+        <field name="bucket" enum-type="bucket_type"/>
+
+        <many-to-one field="message" target-entity="App\Messages\Domain\Entity\Message"/>
+        <many-to-one field="channel" target-entity="App\Channels\Domain\Entity\Channel"/>
+
+        <field name="createdAt" column="created_at" type="datetime_immutable"/>
+        <field name="updatedAt" column="updated_at" type="datetime_immutable" nullable="true"/>
+        <lifecycle-callbacks>
+            <lifecycle-callback type="prePersist" method="setCreatedAt"/>
+        </lifecycle-callbacks>
+        <lifecycle-callbacks>
+            <lifecycle-callback type="preUpdate" method="setUpdatedAt"/>
+        </lifecycle-callbacks>
+    </mapped-superclass>
+</doctrine-mapping>
